+++ packages/extended-yo-generator-test/src/Components/JSONFileMappingTester.ts
@@ -34,14 +34,38 @@ export class JSONFileMappingTester<TGenerator extends IGenerator<TSettings, TOpt
     }
 
     /**
-     * Gets the metadata of inside the file-mapping output.
+     * Gets the parsed content of the source-file.
+     *
+     * @returns
+     * The parsed content of the source-file.
      */
-    public get Metadata(): Promise<any>
+    public async ParseSource(): Promise<any>
     {
-        return (
-            async () =>
-            {
-                return parse(await this.Content);
-            })();
+        return this.Parse(await this.ReadSource());
+    }
+
+    /**
+     * Gets the parsed content of the output-file.
+     *
+     * @returns
+     * The parsed content of the output-file.
+     */
+    public async ParseOutput(): Promise<any>
+    {
+        return this.Parse(await this.ReadOutput());
+    }
+
+    /**
+     * Parses the specified {@link text `text`}.
+     *
+     * @param text
+     * The text to parse.
+     *
+     * @returns
+     * The parsed representation of the specified {@link text `text`}.
+     */
+    public Parse(text: string): any
+    {
+        return parse(text);
     }
 }