+++ packages/extended-yo-generator/src/Components/FileManagement/FileMappingOptions.ts
@@ -4,6 +4,7 @@ import { IGenerator } from "../../IGenerator";
 import { IGeneratorSettings } from "../../IGeneratorSettings";
 import { GeneratorComponent } from "../GeneratorComponent";
 import { FileMapping } from "./FileMapping";
+import { IFileMapping } from "./IFileMapping";
 
 /**
  * Provides data for creating a {@link FileMapping `FileMapping<TSettings, TOptions>`}.
@@ -14,7 +15,7 @@ import { FileMapping } from "./FileMapping";
  * @template TOptions
  * The type of the options of the generator.
  */
-export abstract class FileMappingOptions<TSettings extends IGeneratorSettings, TOptions extends GeneratorOptions> extends GeneratorComponent<TSettings, TOptions, FileMapping<TSettings, TOptions>>
+export abstract class FileMappingOptions<TSettings extends IGeneratorSettings, TOptions extends GeneratorOptions> extends GeneratorComponent<TSettings, TOptions, FileMapping<TSettings, TOptions>> implements IFileMapping<TSettings, TOptions>
 {
     /**
      * Initializes a new instance of the {@link FileMappingOptions `FileMappingOptions<TSettings, TOptions>`} class.
@@ -49,14 +50,6 @@ export abstract class FileMappingOptions<TSettings extends IGeneratorSettings, T
     }
 
     /**
-     * Gets the content of the source-file.
-     */
-    public get Content(): Promise<string>
-    {
-        return this.ReadSource();
-    }
-
-    /**
      * @inheritdoc
      *
      * @returns
@@ -73,6 +66,17 @@ export abstract class FileMappingOptions<TSettings extends IGeneratorSettings, T
     public async Processor?(): Promise<void>;
 
     /**
+     * Reads the contents of the source-file.
+     *
+     * @returns
+     * The contents of the source-file.
+     */
+    public async ReadSource(): Promise<string>
+    {
+        return this.ReadFile(this.Resolved.Source);
+    }
+
+    /**
      * Reads the contents of the file located at the specified {@link path `path`}.
      *
      * @param path
@@ -84,17 +88,6 @@ export abstract class FileMappingOptions<TSettings extends IGeneratorSettings, T
     protected async ReadFile(path: string): Promise<string>
     {
         return (await readFile(path)).toString();
-    }
-
-    /**
-     * Reads the contents of the source-file.
-     *
-     * @returns
-     * The contents of the source-file.
-     */
-    protected async ReadSource(): Promise<string>
-    {
-        return this.ReadFile(this.Resolved.Source);
     }
 
     /**